+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to [lang-feel](https://github.com/nikku/lang-feel) are docum
 ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: turn into pure ES module ([#26](https://github.com/nikku/lang-feel/pull/26))
+* `FEAT`: add FEEL syntax highlighting ([#31](https://github.com/nikku/lang-feel/pull/31))
 * `DEPS`: update to `lezer-feel@2.0.0`
 * `DEPS`: update to `@lezer/common@1.4.0`
 * `DEPS`: update to `@codemirror/autocomplete@6.20.0`